+++ pkg/model/score.go
@@ -1,6 +1,6 @@
 package model
 
 type Score struct {
-	ClientID string `json:"clientId"`
-	Score    int64  `json:"score"`
+	ClientID string  `json:"clientId"`
+	Score    float64 `json:"score"`
 }
